requirements.txt
# === CORE DEPENDENCIES ===
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# === WEB SCRAPING ===
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === DATABASE ===
sqlalchemy>=2.0.0

# === DATA SCIENCE / ML ===
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# === API / WEB SERVER ===
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# === MONITORING / METRICS ===
psutil>=5.9.0

# === VALIDATION ===
pydantic>=2.5.0

# === LOGGING ===
structlog>=23.2.0
colorlog>=6.8.0

# === TESTING (Development) ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# === CODE QUALITY (Development) ===
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# === OPTIONAL: CACHING ===
# redis>=5.0.0

# === OPTIONAL: ADVANCED METRICS ===
# prometheus-client>=0.19.0
